Allow Geo-widget to backoff when enable location is denied
Fixes #231
